Add configure script to make it work on BSD.
We could use `extern const char *__progname` to work on musl/glibc and
BSDs, but it's not a recommended interface, and they prefer
getprogname(). In the interest of proper platform usage, use
getprogname() on BSD and program_invocation_short_name on musl/glibc.

Supporting arc4random_buf and getrandom is pretty simple.

Hasn't been tested on an actual BSD platform.