Clear up input file code in purr.
In order to accept a file from stdin, it is necessary to buffer it in
its entirety, so that the final size is known and can be included in the
HTTP header.

As it stands, the stream mode for mmap_file, when used for reading
files, is not extremely useful.

Also:

- files.c: look at fwrite error.
- mmap_file.c: disable buffering for FILE stream.