Improve gemi's certificate handling.
It still opens certs in ~/.config/gemi, but instead of doing a big fd
dance, simply uses an openat(dirfd(), readdir()->d_name) equivalent.
This way we even save on the extra fd and fcntl() call.

Use fnmatch() to filter for file names that end with ".pem". This can be
done with string manipulation, but fnmatch is easy to extend.